% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCTest-package.R
\docType{package}
\name{GCTest-package}
\alias{GCTest}
\alias{GCTest-package}
\title{GCTest: This was my first complete package (I hope) that can be loaded from GitHub}
\description{
Testing how to build a useful package with data sets
}
\author{
\strong{Maintainer}: Martin Schoch \email{mdschoch@olgchs.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
