% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Data of 500 S&P 500 Stocks
500 rows with 6 columns}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 6 columns.
}
\usage{
sp500
}
\description{
Data of 500 S&P 500 Stocks
500 rows with 6 columns
}
\keyword{datasets}
